fix: R CMD check warnings — C++ format specifier, Rd math notation, .Rbuildignore
- src/functions.cpp: cast size_t args to int in Rprintf to fix -Wformat
- R/Stats.R + man/summaryRuns.Rd: replace bare $F_{ROH}$ with \eqn{F_{ROH}}
- .Rbuildignore: exclude NEWS.html from tarball
- DESCRIPTION: bump Config/roxygen2/version to 8.0.0 (devtools::document side effect)

+++ detectRUNS/src/functions.cpp
@@ -1230,7 +1230,7 @@ DataFrame tableRuns(
     Rprintf(
       "Processing chromosome '%s' (%d/%d)\n",
       chrom.c_str(),
-      i+1, unique_chromosomes.size());
+      (int)(i+1), (int)unique_chromosomes.size());
 
     // extract the desired chrom
     DataFrame runsChrom = subset_runs_by_chrom(runs, chrom);
